Update gems and clean up tests
There was a security issue with redcarpet, as reported by Gemnasium. I went ahead and updated all the gems, and fixed the tests so that they work with the latest version of the demo API.

I also added the rubocop gem and fixed offenses in all modified files.

These changes only affect the development and testing of this gem, so there is no need to bump up the version number.